back_end/api/v1/requirements.txt
# util
tqdm==4.62.2
google-auth==1.35.0
sendgrid==6.1.0
pytest==7.1.2

# scraping
requests==2.23.0
beautifulsoup4==4.9.1

# data analysis
numpy==1.21.2
pandas==1.3.0
pandarallel==1.6.1
joblib==1.0.1

# app
ebaysdk~=2.2.0
fastapi==0.75.2
uvicorn==0.17.6
gunicorn==20.1.0
retrying==1.3.3

# database
firebase-admin==4.4.0
fsspec==0.8.0
gcsfs==0.7.0
google-cloud-bigquery==2.26.0
google-cloud-secret-manager==2.10.0
pyarrow==5.0.0
